refactor: 패키지 구조를 변경한다.
