+++ app/src/main/java/woowacourse/shopping/ShoppingApplication.kt
@@ -1,10 +1,10 @@
 package woowacourse.shopping
 
 import android.app.Application
-import com.boogiwoogi.di.version2.DefaultInstanceContainer
-import com.boogiwoogi.di.version2.DiInjector
-import com.boogiwoogi.di.version2.InstanceContainer
-import com.boogiwoogi.di.version2.Modules
+import com.boogiwoogi.di.DefaultInstanceContainer
+import com.boogiwoogi.di.DiInjector
+import com.boogiwoogi.di.InstanceContainer
+import com.boogiwoogi.di.Modules
 import woowacourse.shopping.di.application.ApplicationModule
 
 class ShoppingApplication : Application() {
