+++ app/src/main/java/woowacourse/shopping/di/activity/ActivityInstanceContainer.kt
@@ -2,9 +2,9 @@ package woowacourse.shopping.di.activity
 
 import androidx.lifecycle.DefaultLifecycleObserver
 import androidx.lifecycle.LifecycleOwner
+import com.boogiwoogi.di.Instance
+import com.boogiwoogi.di.InstanceContainer
 import com.boogiwoogi.di.Qualifier
-import com.boogiwoogi.di.version2.Instance
-import com.boogiwoogi.di.version2.InstanceContainer
 import kotlin.reflect.KClass
 import kotlin.reflect.KParameter
 import kotlin.reflect.full.findAnnotation
