+++ app/src/main/java/woowacourse/shopping/di/activity/ActivityModules.kt
@@ -1,9 +1,9 @@
 package woowacourse.shopping.di.activity
 
 import android.content.Context
+import com.boogiwoogi.di.Modules
 import com.boogiwoogi.di.Provides
 import com.boogiwoogi.di.Qualifier
-import com.boogiwoogi.di.version2.Modules
 import woowacourse.shopping.di.ContextProvider
 import woowacourse.shopping.ui.cart.DateFormatter
 import kotlin.reflect.KClass
