+++ app/src/main/java/woowacourse/shopping/di/application/ApplicationModule.kt
@@ -1,12 +1,12 @@
 package woowacourse.shopping.di.application
 
 import android.content.Context
+import com.boogiwoogi.di.DiInjector
 import com.boogiwoogi.di.Module
+import com.boogiwoogi.di.Modules
 import com.boogiwoogi.di.Provides
 import com.boogiwoogi.di.Qualifier
 import com.boogiwoogi.di.UsableOn
-import com.boogiwoogi.di.version2.DiInjector
-import com.boogiwoogi.di.version2.Modules
 import woowacourse.shopping.data.DatabaseCartRepository
 import woowacourse.shopping.data.DefaultProductRepository
 import woowacourse.shopping.data.InMemoryCartRepository
